# refactor(turbo-tasks): Merge CELL_COUNTERS and CURRENT_TASK_ID into CURRENT_TASK_STATE