--- a/turbopack/crates/turbo-tasks/src/manager.rs
+++ b/turbopack/crates/turbo-tasks/src/manager.rs
@@ -263,6 +263,13 @@ pub struct TurboTasks<B: Backend + 'static> {
 }
 
 struct CurrentTaskState {
+    task_id: TaskId,
+
+    /// A unique identifier created for each unique `CurrentTaskState`. Used to
+    /// check that [`CurrentTaskState::local_cells`] are valid for the current
+    /// `RawVc::LocalCell`.
+    execution_id: ExecutionId,
+
     /// Affected tasks, that are tracked during task execution. These tasks will
     /// be invalidated when the execution finishes or before reading a cell
     /// value.
@@ -270,11 +277,6 @@ struct CurrentTaskState {
 
     /// True if the current task has state in cells
     stateful: bool,
-
-    /// A unique identifier created for each unique `CurrentTaskState`. Used to
-    /// check that [`CurrentTaskState::local_cells`] are valid for the current
-    /// `RawVc::LocalCell`.
-    execution_id: ExecutionId,
 
     /// Tracks how many cells of each type has been allocated so far during this task execution.
     /// When a task is re-executed, the cell count may not match the existing cell vec length.
@@ -288,8 +290,9 @@ struct CurrentTaskState {
 }
 
 impl CurrentTaskState {
-    fn new(execution_id: ExecutionId) -> Self {
+    fn new(task_id: TaskId, execution_id: ExecutionId) -> Self {
         Self {
+            task_id,
             tasks_to_notify: Vec::new(),
             stateful: false,
             execution_id,
@@ -303,8 +306,6 @@ impl CurrentTaskState {
 task_local! {
     /// The current TurboTasks instance
     static TURBO_TASKS: Arc<dyn TurboTasksApi>;
-
-    static CURRENT_TASK_ID: TaskId;
 
     static CURRENT_TASK_STATE: RefCell<CurrentTaskState>;
 }
@@ -520,8 +521,10 @@ impl<B: Backend + 'static> TurboTasks<B> {
         let future = async move {
             let mut schedule_again = true;
             while schedule_again {
-                let task_state =
-                    RefCell::new(CurrentTaskState::new(this.execution_id_factory.get()));
+                let task_state = RefCell::new(CurrentTaskState::new(
+                    task_id,
+                    this.execution_id_factory.get(),
+                ));
                 schedule_again = CURRENT_TASK_STATE
                     .scope(task_state, async {
                         if this.stopped.load(Ordering::Acquire) {
@@ -571,10 +574,7 @@ impl<B: Backend + 'static> TurboTasks<B> {
         };
 
         let future = TURBO_TASKS
-            .scope(
-                self.pin(),
-                CURRENT_TASK_ID.scope(task_id, self.backend.execution_scope(task_id, future)),
-            )
+            .scope(self.pin(), self.backend.execution_scope(task_id, future))
             .in_current_span();
 
         #[cfg(feature = "tokio_tracing")]
@@ -1095,10 +1095,10 @@ impl<B: Backend + 'static> TurboTasksApi for TurboTasks<B> {
         &self,
         f: Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>>,
     ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'static>> {
-        let current_task_id = CURRENT_TASK_ID.get();
         let current_task_state_facade = CURRENT_TASK_STATE.with(|ts| {
             let ts = ts.borrow();
             CurrentTaskState {
+                task_id: ts.task_id,
                 tasks_to_notify: Vec::new(),
                 stateful: false,
                 execution_id: ts.execution_id,
@@ -1106,14 +1106,12 @@ impl<B: Backend + 'static> TurboTasksApi for TurboTasks<B> {
                 local_cells: ts.local_cells.clone(),
             }
         });
+        let current_task_id = current_task_state_facade.task_id;
         Box::pin(TURBO_TASKS.scope(
             turbo_tasks(),
-            CURRENT_TASK_ID.scope(
-                current_task_id,
-                CURRENT_TASK_STATE.scope(
-                    RefCell::new(current_task_state_facade),
-                    self.backend.execution_scope(current_task_id, f),
-                ),
+            CURRENT_TASK_STATE.scope(
+                RefCell::new(current_task_state_facade),
+                self.backend.execution_scope(current_task_id, f),
             ),
         ))
     }
@@ -1227,7 +1225,7 @@ impl<B: Backend + 'static> TaskIdProvider for TurboTasks<B> {
 }
 
 pub(crate) fn current_task(from: &str) -> TaskId {
-    match CURRENT_TASK_ID.try_with(|id| *id) {
+    match CURRENT_TASK_STATE.try_with(|ts| ts.borrow().task_id) {
         Ok(id) => id,
         Err(_) => panic!(
             "{} can only be used in the context of turbo_tasks task execution",
@@ -1429,9 +1427,9 @@ pub fn with_turbo_tasks_for_testing<T>(
 ) -> impl Future<Output = T> {
     TURBO_TASKS.scope(
         tt,
-        CURRENT_TASK_ID.scope(
-            current_task,
-            CURRENT_TASK_STATE.scope(RefCell::new(CurrentTaskState::new(execution_id)), f),
+        CURRENT_TASK_STATE.scope(
+            RefCell::new(CurrentTaskState::new(current_task, execution_id)),
+            f,
         ),
     )
 }
@@ -1445,7 +1443,7 @@ pub fn spawn_detached_for_testing(f: impl Future<Output = Result<()>> + Send + '
 }
 
 pub fn current_task_for_testing() -> TaskId {
-    CURRENT_TASK_ID.with(|id| *id)
+    CURRENT_TASK_STATE.with(|ts| ts.borrow().task_id)
 }
 
 /// Get an [`Invalidator`] that can be used to invalidate the current task